fix: add 'px' to zero length value in combobox token
This length value token is currently '0', instead of '0px', which throws
off some CSS calc() results, making them invalid (or requires
multiplying this token by 1px).

This fix makes this value consistent with the rest of the token values,
which all specify 0 length values as '0px'. The only other '0' within
tokens is for an opacity (which is okay).